1. Split CosmosDiagnosticContex to a separate file. 2. Added getContactedRegions method to CosmosDagnostics.
